Support killing active commands for resolved alerts
By default a command matching a resolved alert will be killed.

When a 'fired' alert is received, the message is checked for alert fingerprints that match the command. If one is found, the command will listen for a 'resolved' condition for that fingerprint while its running.

When a 'resolved' alert is received, active commands that match the fingerprint will be killed.

If you don't want a command to stop if a matching alert is resolved while its running, set the `ignore_resolved: true` field in the yaml config for a command.

By default the sender of alerts is notified if a command triggered by the alert failed, by way of an HTTP 500 response. Alertmanager may consider this to be a 'failure to notify', and could re-send the alert.

If you don't want this behaviour, set `notify_on_failure: false` in the yaml config for a command.

By default there is no limit to the number of active commands that could be running for matching alerts.

If you want to set a limit for this, set the `max` fielf in the yaml config for a command.

* Refactored code. I think it's more organized and has better separation of concerns than before.
* Some new modules to help with resolved alert and max commands functionality.
* Server instances are fully separate now; Can run multiple Servers in the same program if one wanted to.